add all the inventory test from ApiGatewayControllerTest to v2/InventoryControllerTest(All tests are now there)
